+++ dbx/callbacks.py
@@ -1,11 +1,14 @@
+import json
 from http import client as http_client
+from json import JSONDecodeError
 from pathlib import Path
 from typing import Optional
 
 import typer
-from dbx.utils import dbx_echo
 
 from dbx import __version__
+from dbx.models.parameters import ExecuteWorkloadParamInfo
+from dbx.utils import dbx_echo
 
 
 def verify_jinja_variables_file(_, value: Optional[Path]) -> Optional[Path]:
@@ -57,3 +60,16 @@ def debug_callback(_, value):
     if value:
         dbx_echo(":bug: Debugging mode is [red]on[/red]")
         http_client.HTTPConnection.debuglevel = 1
+
+
+def execute_parameters_callback(_, value: str) -> Optional[str]:
+    if value:
+        try:
+            _parsed = json.loads(value)
+        except JSONDecodeError as e:
+            dbx_echo(":boom: Provided parameters payload cannot be parsed since it's not in json format")
+            raise e
+
+        ExecuteWorkloadParamInfo(**_parsed)
+
+        return value