+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ✨Add `--workflows` argument to `dbx deploy`
 - ✨Add `--assets-only` and `--from-assets` as a clearer replacement for old arguments
 - ⚡️Add support for `--environment` parameter for `dbx sync` commands
+- ✨Add flexible parameter overriding logic for `dbx execute` via new `--parameters` option
+
 
 ## Changed
 - ✨Arguments `--allow-delete-unmatched`/`--disallow-delete-unmatched` were **replaced** with `--unmatched-behaviour` option.